Move S3[Client|Config]Module to the new storage lib
