+++ catalogue_pipeline/recorder/src/main/scala/uk/ac/wellcome/platform/recorder/Server.scala
@@ -8,6 +8,7 @@ import uk.ac.wellcome.finatra.controllers.ManagementController
 import uk.ac.wellcome.finatra.modules._
 import uk.ac.wellcome.messaging.sqs.{SQSClientModule, SQSConfigModule}
 import uk.ac.wellcome.storage.dynamo.{DynamoClientModule, DynamoConfigModule}
+import uk.ac.wellcome.storage.s3.{S3ClientModule, S3ConfigModule}
 import uk.ac.wellcome.platform.recorder.modules.RecorderModule
 
 object ServerMain extends Server
