+++ reindexer/reindex_worker/src/main/scala/uk/ac/wellcome/platform/reindex_worker/Server.scala
@@ -14,6 +14,7 @@ import uk.ac.wellcome.messaging.sns.{SNSClientModule, SNSConfigModule}
 import uk.ac.wellcome.messaging.sqs.{SQSClientModule, SQSConfigModule}
 import uk.ac.wellcome.platform.reindex_worker.modules.ReindexerWorkerModule
 import uk.ac.wellcome.storage.dynamo.{DynamoClientModule, DynamoConfigModule}
+import uk.ac.wellcome.storage.s3.{S3ClientModule, S3ConfigModule}
 
 object ServerMain extends Server
