+++ sbt_common/storage/src/main/scala/uk/ac/wellcome/storage/s3/S3ClientModule.scala
@@ -1,4 +1,4 @@
-package uk.ac.wellcome.finatra.modules
+package uk.ac.wellcome.storage.s3
 
 import javax.inject.Singleton
